Rename files to avoid self-inclusion when bundling with Webpack and RollUp Fixes #89
