+++ package.json
@@ -2,11 +2,11 @@
   "name": "ng2-nouislider",
   "version": "1.7.8",
   "description": "Angular2 noUiSlider component",
-  "main": "src/nouislider.js",
-  "typings": "src/nouislider.d.ts",
+  "main": "src/ng2-nouislider.js",
+  "typings": "src/ng2-nouislider.d.ts",
   "scripts": {
     "contributors": "all-contributors generate",
-    "lint": "tslint src/nouislider.ts",
+    "lint": "tslint src/ng2-nouislider.ts",
     "build": "npm run lint && ngc",
     "start": "npm run build && webpack-dev-server --host 0.0.0.0",
     "deploy": "NODE_ENV=prod npm run build && NODE_ENV=prod webpack && gh-pages -d demo",
