+++ src/ng2-nouislider.d.ts
