+++ src/ng2-nouislider.js
