+++ src/ng2-nouislider.metadata.json
