+++ src/ng2-nouislider.ts
