+++ tsconfig.json
@@ -25,7 +25,7 @@
     "node_modules"
   ],
   "files": [
-    "src/nouislider.ts"
+    "src/ng2-nouislider.ts"
   ],
   "angularCompilerOptions": {
     "genDir": "compiled"
